further reorganization, all scripts working
